+++ database/migrations/2022_02_08_152347_create_endereco.php
@@ -21,8 +21,8 @@ class CreateEndereco extends Migration
             $table->string('pais')->nullable();
             $table->string('bairro')->nullable();
             $table->string('numero')->nullable();
-            $table->integer('estado_id');
-            $table->integer('municipio_id');
+            $table->unsignedBigInteger('estado_id');
+            $table->unsignedBigInteger('municipio_id');
             $table->boolean('ativo')->default(true);
 
             $table->foreign('estado_id')->references('id')->on('estado');
