+++ database/migrations/2022_02_09_484041_create_papel.php
@@ -15,7 +15,7 @@ class CreatePapel extends Migration
     {
         Schema::create('papel', function (Blueprint $table) {
             $table->id();
-            $table->integer('permissao_id');
+            $table->unsignedBigInteger('permissao_id');
             $table->string('acao');
             $table->string('descricao');
             $table->boolean('ativo')->default(true);
