+++ database/migrations/2022_02_09_584041_create_menu.php
@@ -19,7 +19,7 @@ class CreateMenu extends Migration
             $table->integer('nivel');
             $table->string('icone');
             $table->string('icone_aux');
-            $table->integer('papel_id');
+            $table->unsignedBigInteger('papel_id');
             $table->string('rota')->nullable();
             $table->string('menu_pai_id')->nullable();
             $table->boolean('ativo')->default(true);
