+++ database/migrations/2022_02_09_784041_create_perfil_papel.php
@@ -15,8 +15,8 @@ class CreatePerfilPapel extends Migration
     {
         Schema::create('perfil_papel', function (Blueprint $table) {
             $table->id();
-            $table->integer('perfil_id');
-            $table->integer('papel_id');
+            $table->unsignedBigInteger('perfil_id');
+            $table->unsignedBigInteger('papel_id');
             $table->timestamps();
 
             $table->foreign('perfil_id')->references('id')->on('perfil');
