+++ database/migrations/2022_02_14_192244_create_ordem_servico_faturamento.php
@@ -15,7 +15,7 @@ class CreateOrdemServicoFaturamento extends Migration
     {
         Schema::create('ordem_servico_faturamento', function (Blueprint $table) {
             $table->id();
-            $table->integer('ordem_servico_id');
+            $table->unsignedBigInteger('ordem_servico_id');
             $table->boolean('ativo')->default(true);
             $table->timestamps();
